Automated clean up and removal devices that are not responding or orphan devices. (#768)
Changes

Major improvement in the handling of missing or orphan devices in Alexa. When deviceCleanup is enabled the plugin will inform Alexa when a request is received for a missing or orphan device, and tell Alexa to remove the device.
Bug Fix

Fix for Routines not working Alexa Routines no longer working #761
Fix for missing Mute function with Apple TV
Changes

Increased test case coverage